Created the routes for firebase, reminder to edit your .env file according to the config.py files.
