+++ backend/routes/prediction_routes.py
@@ -1,0 +1,97 @@
+from datetime import datetime
+from flask import Blueprint, request, jsonify # type: ignore
+import firebase_admin # type: ignore
+from firebase_admin import credentials,firestore # type: ignore
+from config import Config
+
+firebase_config_json = Config.FIREBASE_CONFIG
+cred = credentials.Certificate(firebase_config_json)
+
+db = firestore.client()
+prediction_bp = Blueprint('prediction_bp', __name__)
+
+## Adding patient details    
+@prediction_bp.route('/add_details', methods=['POST'])
+def add_patient_details():
+    try:
+        data = request.json
+        details_ref = db.collection('details').document(data['detailId'])
+        details_ref.set({
+            'patientId': data['patientId'],
+            'timestamp': datetime.utcnow(),
+            'age': data['age'],
+            'gender': data['gender'],
+            'weight': data['weight'],
+            'height': data['height'],
+            'bmi': data['bmi'],
+            'bloodPressure': data['bloodPressure'],
+            'heartRate': data['heartRate'],
+            'smokingStatus': data['smokingStatus'],
+            'alcoholConsumption': data['alcoholConsumption'],
+            'dietaryHabits': data['dietaryHabits'],
+            'exerciseFrequency': data['exerciseFrequency'],
+            'existingConditions': data['existingConditions'],
+            'medications': data['medications'],
+            'surgicalHistory': data['surgicalHistory'],
+            'familyMedicalHistory': data['familyMedicalHistory'],
+            'birthWeight': data['birthWeight'],
+            'gestationalAge': data['gestationalAge'],
+            'maternalHealthComplications': data['maternalHealthComplications'],
+            'multiplePregnancy': data['multiplePregnancy'],
+            'congenitalAnomalies': data['congenitalAnomalies'],
+            'pretermBirth': data['pretermBirth'],
+            'lowBirthWeight': data['lowBirthWeight']
+        })
+        return jsonify({"message": "Patient details added successfully"}), 201
+
+    except Exception as e:
+        return jsonify({"error": str(e)}), 500
+
+## Add prediction for patient
+@prediction_bp.route('/add_prediction', methods=['POST'])
+def add_prediction():
+    try:
+        data = request.json
+        prediction_ref = db.collection('predictions').document(data['predictionId'])
+        prediction_ref.set({
+            'patientId': data['patientId'],
+            'doctorId': data['doctorId'],
+            'detailId' : data['detailId'],
+            'predictionResult': data['predictionResult'],
+            'confidenceScore': data['confidenceScore'],
+            'timestamp': datetime.utcnow()
+        })
+
+        contributing_factors_ref = prediction_ref.collection('contributingFactors')
+        for factor, details in data['contributingFactors'].items():
+            contributing_factors_ref.document(factor).set({
+                'factor': factor,
+                'description': details['desc'],
+                'contributionPercentage': details['percentage']
+        })
+        
+        return jsonify({"message": "Prediction added successfully"}), 201
+    except Exception as e:
+        return jsonify({"error": str(e)}), 400
+
+## Fetch prediction by predictionID
+@prediction_bp.route('/get_prediction/<PredictionID>', methods=['GET'])
+def get_prediction(PredictionID):
+    prediction_ref = db.collection('predictions').document(PredictionID)
+    prediction = prediction_ref.get()
+
+    if prediction.exists:
+        return jsonify(prediction.to_dict()), 200
+    else:
+        return jsonify({"error": "Prediction not found"}), 404
+
+## Fetch all predictions for a certain patient
+@prediction_bp.route('/get_all_predictions/<PatientID>', methods=['GET'])
+def get_predictions(PatientID):
+    predictions_ref = db.collection('predictions').where("patientId", "==", PatientID).stream()
+    predictions = [doc.to_dict() for doc in predictions_ref]
+
+    if predictions:
+        return jsonify(predictions), 200
+    else:
+        return jsonify({"message": "No predictions found for this patient"}), 404
